reorganise integration tests
