+++ packages/graphql/tests/integration/filtering/composite-where.int.test.ts
@@ -18,8 +18,8 @@
  */
 
 import { generate } from "randomstring";
-import type { UniqueType } from "../utils/graphql-types";
-import { TestHelper } from "../utils/tests-helper";
+import type { UniqueType } from "../../utils/graphql-types";
+import { TestHelper } from "../../utils/tests-helper";
 
 describe("composite-where", () => {
     const testHelper = new TestHelper();
