+++ packages/graphql/tests/integration/filtering/field-filtering.int.test.ts
@@ -19,7 +19,7 @@
 
 import { gql } from "graphql-tag";
 import { generate } from "randomstring";
-import { TestHelper } from "../utils/tests-helper";
+import { TestHelper } from "../../utils/tests-helper";
 
 describe("field-filtering", () => {
     const testHelper = new TestHelper();
